requirements.txt
tensorflow==2.14.0
absl-py
tqdm
numpy
scipy
tensorflow_addons
Pillow
